refactor: 패키지 구조 변경
- app 패키지에서 service 패키지로 SwaggerConfig, SwaggerOAuthPropertyDecryptor, TestRedisController, TestMongoController, AppApplication, HealthCheckController 클래스 이동
- AppApplication 클래스 이름을 Application으로 변경
- Application 클래스에서 ComponentScan 어노테이션 제거
