+++ service-api/src/main/kotlin/com/millo/ollim/Application.kt
@@ -1,17 +1,15 @@
-package com.millo.ollim.app
+package com.millo.ollim
 
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.autoconfigure.domain.EntityScan
 import org.springframework.boot.runApplication
-import org.springframework.context.annotation.ComponentScan
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories
 
 @SpringBootApplication
 @EnableJpaRepositories(basePackages = ["com.millo.ollim.core.domain"])
 @EntityScan(basePackages = ["com.millo.ollim.core.domain"])
-@ComponentScan(basePackages = ["com.millo.ollim"])
-class AppApplication
+class Application
 
 fun main(args: Array<String>) {
-    runApplication<AppApplication>(*args)
+    runApplication<Application>(*args)
 }
