+++ service-api/src/main/kotlin/com/millo/ollim/HealthCheckController.kt
@@ -1,4 +1,4 @@
-package com.millo.ollim.app
+package com.millo.ollim
 
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.tags.Tag
