+++ service-api/src/main/kotlin/com/millo/ollim/config/SwaggerConfig.kt
@@ -1,4 +1,4 @@
-package com.millo.ollim.app.config
+package com.millo.ollim.config
 
 import io.swagger.v3.oas.models.Components
 import io.swagger.v3.oas.models.OpenAPI
